+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/Misc/EffectInsaneGravity.lua
@@ -3,16 +3,62 @@
 ---@field initialX number
 ---@field initialY number
 ---@field initialZ number
+---@field fallTriggered boolean
 
 ---@class EffectInsaneGravity : ChaosEffectBase
 ---@field initialPlayerX number
 ---@field initialPlayerY number
 ---@field initialPlayerZ number
 ---@field affectedZombies table<IsoZombie, EffectInsaneGravityZombieData>
+---@field playerFallTriggered boolean
 EffectInsaneGravity = ChaosEffectBase:derive("EffectInsaneGravity", "insane_gravity")
 
 local RADIUS = 75
 local VEHICLE_DOWN_IMPULSE = 100000 * 5 * 2
+local PLAYER_REANIMATE_TIMER = 1
+local ZOMBIE_REANIMATE_TIMER = 10
+
+local function IsPlayerDownState(player)
+    return player:isCurrentState(PlayerFallDownState.instance())
+        or player:isCurrentState(PlayerOnGroundState.instance())
+        or player:isCurrentState(PlayerKnockedDown.instance())
+end
+
+local function TriggerPlayerFall(player)
+    if not player then return end
+
+    player:setFallOnFront(true)
+    player:setKnockedDown(true)
+    player:setOnFloor(false)
+    player:setReanimateTimer(PLAYER_REANIMATE_TIMER)
+    player:setIgnoreMovement(true)
+    player:setBlockMovement(true)
+    player:changeState(PlayerFallDownState.instance())
+end
+
+local function MaintainPlayerDown(player)
+    if not player then return end
+
+    player:setReanimateTimer(PLAYER_REANIMATE_TIMER)
+    player:setIgnoreMovement(true)
+    player:setBlockMovement(true)
+end
+
+local function TriggerZombieFall(zombie)
+    if not zombie or zombie:isDead() then return end
+
+    zombie:setFallOnFront(true)
+    zombie:setKnockedDown(true)
+    zombie:setOnFloor(false)
+    zombie:setReanimateTimer(ZOMBIE_REANIMATE_TIMER)
+    zombie:changeState(ZombieFallDownState.instance())
+end
+
+local function MaintainZombieDown(zombie)
+    if not zombie or zombie:isDead() then return end
+
+    zombie:setReanimateTimer(ZOMBIE_REANIMATE_TIMER)
+end
 
 function EffectInsaneGravity:OnStart()
     ChaosEffectBase:OnStart()
@@ -26,6 +72,7 @@ function EffectInsaneGravity:OnStart()
     self.initialPlayerY = player:getY()
     self.initialPlayerZ = player:getZ()
     self.affectedZombies = {}
+    self.playerFallTriggered = false
 end
 
 ---@param deltaMs integer
@@ -55,7 +102,17 @@ function EffectInsaneGravity:OnTick(deltaMs)
     player:setX(self.initialPlayerX)
     player:setY(self.initialPlayerY)
     player:setZ(self.initialPlayerZ)
-    player:setKnockedDown(true)
+
+    if not self.playerFallTriggered then
+        TriggerPlayerFall(player)
+        self.playerFallTriggered = true
+    elseif not IsPlayerDownState(player) then
+        player:setFallOnFront(true)
+        player:setOnFloor(true)
+        player:changeState(PlayerOnGroundState.instance())
+    end
+
+    MaintainPlayerDown(player)
 
     local affectedZombies = self.affectedZombies
     ChaosZombie.ForEachZombieInRange(self.initialPlayerX, self.initialPlayerY, RADIUS, function(zombie)
@@ -68,25 +125,58 @@ function EffectInsaneGravity:OnTick(deltaMs)
                 initialX = zombie:getX(),
                 initialY = zombie:getY(),
                 initialZ = zombie:getZ(),
+                fallTriggered = false,
             }
             affectedZombies[zombie] = data
-            zombie:setUseless(true)
+
+            local npc = ChaosNPCUtils.GetNPCFromZombie(zombie)
+            if npc then
+                npc:AddDisableAiEffect("insane_gravity")
+            end
         end
 
         zombie:setX(data.initialX)
         zombie:setY(data.initialY)
         zombie:setZ(data.initialZ)
-        zombie:setKnockedDown(true)
-    end, true, nil)
+
+        if not data.fallTriggered then
+            TriggerZombieFall(zombie)
+            data.fallTriggered = true
+            zombie:setUseless(true)
+        elseif not zombie:isCurrentState(ZombieFallDownState.instance()) and not zombie:isCurrentState(ZombieOnGroundState.instance()) then
+            zombie:setFallOnFront(true)
+            zombie:setOnFloor(true)
+            zombie:changeState(ZombieOnGroundState.instance())
+        end
+
+        MaintainZombieDown(zombie)
+    end, false, nil)
 end
 
 function EffectInsaneGravity:OnEnd()
     ChaosEffectBase:OnEnd()
 
+    local player = getPlayer()
+    if player then
+        player:setIgnoreMovement(false)
+        player:setBlockMovement(false)
+        player:setBumpType("")
+        player:setBumpDone(true)
+
+        if player:isCurrentState(PlayerOnGroundState.instance()) then
+            player:changeState(PlayerGetUpState.instance())
+        end
+    end
+
     if self.affectedZombies then
         for zombie, data in pairs(self.affectedZombies) do
             if zombie and zombie:isAlive() then
                 zombie:setUseless(data.wasUseless)
+                zombie:setReanimateTimer(0)
+                local npc = ChaosNPCUtils.GetNPCFromZombie(zombie)
+                if npc then
+                    npc:RemoveDisableAiEffect("insane_gravity")
+                end
             end
         end
         self.affectedZombies = {}